password-strength-meter/pyproject.toml
[project]
name = "password-strength-meter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyperclip>=1.9.0",
    "zxcvbn-python>=4.4.24",
]
